API を実装
