+++ src/server/app/Http/Presenters/Api/Creator/DeletePresenter.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http\Presenters\Api\Creator;
+
+use Illuminate\Http\JsonResponse;
+
+class DeletePresenter
+{
+    public function present(): JsonResponse
+    {
+        return response()->json(status: 204);
+    }
+}
